UI/revise UI manualy (#55)
* feat(database): add new database versions 11 and 12 with updated schemas for sakes and reviews

feat(sake): implement SakeListSection composable for displaying sake items in expandable sections

test(review): add unit tests for ReviewMode to verify soundness items and debug mode functionality

* some minor UI fix (spacing)

* 未参照アプリ内画像の自動削除をデフォルトにし、設定自体を削除

* feat(sake): refactor section handling in SakeList to use state management for expanded sections

* Fix androidTest expectation values for fixed functions

+++ app/src/main/java/io/github/pyth0n14n/tastinggenie/data/local/entity/SakeEntity.kt
@@ -20,7 +20,7 @@ data class SakeEntity(
     val maker: String?,
     val prefecture: Prefecture?,
     val city: String? = null,
-    val alcohol: Int?,
+    val alcohol: Float?,
     val kojiMai: String?,
     val kojiPolish: Int?,
     val kakeMai: String?,
